refactor to a object
